+++ server/src/provider.js
@@ -613,7 +613,7 @@ Provider.prototype.parseHarbour = function (words) {
             this.currentMethod = undefined;
             this.currentClass.endLine = this.lineNr;
         } else if (words[0].length >= 4) {
-            if ((words[0] == "class") || (words[0] == "create" && words[1] == "class")) {
+            if (((words[0] == "class") && !this.currentClass) || (words[0] == "create" && words[1] == "class")) {
                 if (this.currentMethod) this.currentMethod.endLine = this.lastCodeLine;
                 this.currentMethod = undefined;
                 if (words[0] == "create")
@@ -622,13 +622,14 @@ Provider.prototype.parseHarbour = function (words) {
                     this.currentClass = this.addInfo(words1, 'class', "definition")
                 this.currentClass.endLine = undefined;
             } else
-                if (words[0] == "data" || words[0] == "var") {
+                if(words[0] == "data" || words[0] == "var" || words[0] == "classdata" ||
+                    words[0] == "classvar" || (words[0] == "class" && words[1]=="var"))  {
                     if (this.currentClass) {
                         words[1] = words1;
                         this.parseDeclareList(words.slice(1).join(" "), 'data', this.currentClass)
                     }
                 } else
-                    if (words[0] == "method".substring(0, words[0].length)) {
+                    if (words[0] == "method" || words[0] == "classmethod" || (words[0] == "class" && words[1]=="method")) {
                         var r = methodRegEx.exec(this.currLine);
                         if (r) {
                             var fLike = "definition"